fix ace/PowerPacker bug
Reading some .ace files (or any other format packed with PowerPack),
it sometimes reads invalid offsets while unpacking.

This was caused by an error in ReverseBitReader.eof?, which sometimes
returned _true_ when at the end of the bit stream, causing the depack
routine to continue depacking (and eventually yielding a wrong offset).